Extract chat discussion round state
