+++ dev-docs/architecture.md
@@ -66,6 +66,7 @@ js/
   chat-marker-prompts.js — Per-marker and selected-correlation AI prompt builders
   chat-continuation.js — Response limit detection and automatic continuation
   chat-discussion.js — Multi-persona discussion/debate orchestration
+  chat-discussion-round-state.js — Multi-persona discussion round persistence helpers
   chat-discussion-state.js — Multi-persona discussion persona/thread state helpers
   chat-discussion-ui.js — Multi-persona discussion picker and continuation controls
   chat-empty-state.js — Empty chat and onboarding message renderer
