+++ tests/test-correctness-phase2.js
@@ -135,6 +135,7 @@ const pwaAppShellAssets = [
   '/js/chat-nudge.js',
   '/js/chat-panel.js',
   '/js/chat-discussion.js',
+  '/js/chat-discussion-round-state.js',
   '/js/chat-discussion-state.js',
   '/js/chat-discussion-ui.js',
   '/js/chat-onboarding.js',
